Automate missing data polishing for tsibble

It is an iterative process for minimising the loss until a tolerance value.

na_polish_auto() returns the polished data.

na_polish_auto_trace() returns a metric tibble for documenting the steps.

na_polish_auto(data, cutoff, tol = 0.1, funs = na_polish_funs(),
  quiet = FALSE)

na_polish_auto_trace(data, cutoff, tol = 0.1, funs = na_polish_funs(),
  quiet = FALSE)

Arguments

data
A tsibble.

cutoff
A numeric between 0 and 1. Rows/cols will be kept, if the proportion of overall missings is less than the cutoff.

tol
A tolerance value close or equal to zero as stopping rule. It compares to the loss defined as prop_na * prop_removed to be minimised. See na_polish_metrics() for details.

funs
A list of na_polish_*() functions to go through.

quiet
If FALSE, report metrics while automatically polishing, and requires the "cliapp" package to be installed.

Examples

wdi_ts <- tsibble::as_tsibble(wdi, key = country_code, index = year)
wdi_after <- na_polish_auto(wdi_ts, cutoff = .8)

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

#> 

#> ── Pass 1 ──────────────────────────────────────────────────────────────────────

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

#> 1. `na_polish_index()` 0.673 * 0.335 = 0.225

#> 2. `na_polish_measures()` 0.881 * 0.109 = 0.096

#> 3. `na_polish_index2()` 0.570 * 0.113 = 0.064

#> 4. `na_polish_key()` 0.837 * 0.000 = 0.000

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

#> ✔ 0.328

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610
#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

#> 

#> ── Pass 2 ──────────────────────────────────────────────────────────────────────

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

#> 1. `na_polish_index()` 0.432 * 0.157 = 0.068

#> 2. `na_polish_index2()` 0.256 * 0.055 = 0.014

#> 3. `na_polish_key()` 0.816 * 0.000 = 0.000

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

#> ✔ 0.080

#> Error in processx_conn_write(con, str, sep, encoding) : 
#>   Cannot write connection: Bad file descriptor at processx-connection.c:610

na_polish_metrics(wdi_ts, wdi_after)

#> # A tibble: 1 x 6
#>   prop_na nobs_na prop_removed nobs_removed nrows_removed ncols_removed
#>     <dbl>   <int>        <dbl>        <int>         <int>         <int>
#> 1   0.636  220533        0.581       346948          5752             6